[new] Make 'depends' function behave depending on the stereotype
If simple stereotype is given (not ending with '<' or '>'), then no geometry
indications are included. If the stereotype ends with '>' or '<', then a
geometry indication for PlantUML is included:

  client ..> supplier : stereotype
  client -[hidden]right- supplier

That makes that, e.g.,

  depends("User", "Touch_Screen", "interacts >")  '(1)
  depends("User", "Touch_Screen", "interacts")    '(2)

case (1) draws "User" and "Touch_Screen" at the same level, whilst case (2)
draws "User" on top of "Touch_Screen".